Add CHANGELOG, test CI workflow, ruff cleanup
- ``CHANGELOG.md`` — single big entry for ``1.0.0`` covering the full
  tool surface (~125 tools, 17 modules), backends + transports,
  runtime version-gating contract (``required_version`` /
  ``actual_version`` payload), opt-in AppImage auto-fetch, helper
  scripts, and the bug fixes that landed before the tag (16 MiB stream
  buffer, ``pgrep -x`` dup-guard, wheel build ``force-include``).
- ``.github/workflows/test.yml`` — runs unit tests across
  ubuntu/windows/macos × Python 3.11/3.12/3.13, plus ``ruff check``
  and a ``uv build`` job that pipes through ``twine check`` and
  smoke-installs the wheel into a fresh venv. Live (Phase 1 / Phase 2)
  tests skip cleanly without ``SCRIBUS_MCP_LIVE`` since they need a
  real Scribus install.
- Ruff cleanup so the new lint job is green from day one: ``Callable``
  imports moved to ``collections.abc`` (UP035); two ``__all__`` lists
  isort-sorted (RUF022); unused ``json`` import removed (F401);
  redundant ``int(round(...))`` calls dropped (RUF046, 9 sites);
  string match patterns made raw / nested ``with`` collapsed in
  appimage tests (RUF043, SIM117); a stray quoted type annotation and
  one unsorted import block fixed (UP037, I001). 19 errors → 0.

+++ src/scribus_mcp/tools/text.py
@@ -224,8 +224,8 @@ def register(mcp, ctx: ServerCtx) -> None:
         1/100 pt).
         """
         backend = await get_backend(ctx, mode)
-        offset_hpt = -1 if offset_pt == -1.0 else int(round(offset_pt * 100))
-        width_hpt = -1 if width_pt == -1.0 else int(round(width_pt * 100))
+        offset_hpt = -1 if offset_pt == -1.0 else round(offset_pt * 100)
+        width_hpt = -1 if width_pt == -1.0 else round(width_pt * 100)
         result = await backend.call("setUnderline", offset_hpt, width_hpt, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
 
@@ -261,8 +261,8 @@ def register(mcp, ctx: ServerCtx) -> None:
         conversion as ``set_underline_pt`` (× 100 → Scribus's 1/100 pt).
         """
         backend = await get_backend(ctx, mode)
-        offset_hpt = -1 if offset_pt == -1.0 else int(round(offset_pt * 100))
-        width_hpt = -1 if width_pt == -1.0 else int(round(width_pt * 100))
+        offset_hpt = -1 if offset_pt == -1.0 else round(offset_pt * 100)
+        width_hpt = -1 if width_pt == -1.0 else round(width_pt * 100)
         result = await backend.call("setStrikethru", offset_hpt, width_hpt, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
 
@@ -286,7 +286,7 @@ def register(mcp, ctx: ServerCtx) -> None:
         if width_pt < 0:
             return {"ok": False, "error": "width_pt must be >= 0"}
         backend = await get_backend(ctx, mode)
-        width_hpt = int(round(width_pt * 100))
+        width_hpt = round(width_pt * 100)
         result = await backend.call("setOutline", width_hpt, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
 
@@ -322,8 +322,8 @@ def register(mcp, ctx: ServerCtx) -> None:
         ``x_offset_pt`` moves it right.
         """
         backend = await get_backend(ctx, mode)
-        x_hpt = int(round(x_offset_pt * 100))
-        y_hpt = int(round(y_offset_pt * 100))
+        x_hpt = round(x_offset_pt * 100)
+        y_hpt = round(y_offset_pt * 100)
         result = await backend.call("setShadow", x_hpt, y_hpt, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
 
@@ -341,7 +341,7 @@ def register(mcp, ctx: ServerCtx) -> None:
         if scale_percent <= 0:
             return {"ok": False, "error": "scale_percent must be > 0"}
         backend = await get_backend(ctx, mode)
-        scribus_value = int(round(scale_percent * 10))
+        scribus_value = round(scale_percent * 10)
         result = await backend.call("setTextScalingH", scribus_value, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
 
@@ -358,7 +358,7 @@ def register(mcp, ctx: ServerCtx) -> None:
         if scale_percent <= 0:
             return {"ok": False, "error": "scale_percent must be > 0"}
         backend = await get_backend(ctx, mode)
-        scribus_value = int(round(scale_percent * 10))
+        scribus_value = round(scale_percent * 10)
         result = await backend.call("setTextScalingV", scribus_value, name)
         return {"ok": result.ok, "value": result.unwrap_or(), "error": result.error}
